pyproject.toml
[project]
name = "mail-assistant"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
license = {file = "LICENSE"}
dependencies = [
    "pandas",
    "imapclient",
    "duckdb>=1.2.1",
    "python-dotenv>=1.0.1",
    "pyarrow>=19.0.1",
    "fastparquet>=2024.11.0",
    "tqdm>=4.67.1",
    "rich>=13.9.4",
    "schedule>=1.2.2",
    "chromadb>=0.6.3",
    "pydantic-ai>=0.0.44",
    "mcp[cli]>=1.5.0",
]

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
pythonpath = [
  "."
]

[dependency-groups]
dev = [
    "pytest>=8.3.5",
]
